SICAME/requirements.txt
astroid==2.2.5
cairocffi==1.1.0
CairoSVG==2.4.1
cffi==1.12.3
colorama==0.4.1
cssselect2==0.2.2
defusedxml==0.6.0
Django==2.2.5
django-appconf==1.0.3
django-autocomplete-light==3.4.1
django-betterforms==1.2
django-ckeditor==5.7.1
django-easy-pdf==0.1.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-smart-selects==1.5.4
html5lib==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pilkit==2.0
Pillow==6.1.0
psycopg2==2.8.3
pycparser==2.19
pylint==2.3.1
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2019.1
reportlab==3.5.23
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
tinycss2==1.0.2
typed-ast==1.4.0
WeasyPrint==48
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
